+++ src/eca/config.clj
@@ -76,13 +76,17 @@
                                           "gpt-4.1" {}
                                           "gpt-4o" {}
                                           "grok-code-fast-1" {}
-                                          "gemini-2.5-pro" {}}}
+                                          "gemini-2.5-pro" {}
+                                          "gemini-3-pro-preview" {}
+                                          "gemini-3-flash-preview" {}}}
                "google" {:api "openai-chat"
                          :url "${env:GOOGLE_API_URL:https://generativelanguage.googleapis.com/v1beta/openai}"
                          :key "${env:GOOGLE_API_KEY}"
                          :requiresAuth? true
                          :models {"gemini-2.0-flash" {}
-                                  "gemini-2.5-pro" {}}}
+                                  "gemini-2.5-pro" {}
+                                  "gemini-3-pro-preview" {}
+                                  "gemini-3-flash-preview" {}}}
                "ollama" {:url "${env:OLLAMA_API_URL:http://localhost:11434}"}}
    :defaultBehavior "agent"
    :behavior {"agent" {:systemPrompt "${classpath:prompts/agent_behavior.md}"
@@ -301,8 +305,7 @@
                           (map? m*)
                           (let [apply-kebab-key? (applies? kc-paths cur-path)
                                 apply-string-key? (applies? str-paths cur-path)
-                                apply-keywordize-val? (applies? keywordize-paths cur-path)
-                                ]
+                                apply-keywordize-val? (applies? keywordize-paths cur-path)]
                             (into {}
                                   (map (fn [[k v]]
                                          (let [base-name (cond